Moved from double to float in BaseCollisionChecker.

+++ src/org/anddev/andengine/collision/BaseCollisionChecker.java
@@ -36,7 +36,7 @@ public class BaseCollisionChecker {
 				pTopB < pBottomA);
 	}
 
-	public static boolean checkLineCollision(final double pX1, final double pY1, final double pX2, final double pY2, final double pX3, final double pY3, final double pX4, final double pY4) {
+	public static boolean checkLineCollision(final float pX1, final float pY1, final float pX2, final float pY2, final float pX3, final float pY3, final float pX4, final float pY4) {
 		return ((relativeCCW(pX1, pY1, pX2, pY2, pX3, pY3) * relativeCCW(pX1, pY1, pX2, pY2, pX4, pY4) <= 0)
 				&& (relativeCCW(pX3, pY3, pX4, pY4, pX1, pY1) * relativeCCW(pX3, pY3, pX4, pY4, pX2, pY2) <= 0));
 	}
@@ -78,14 +78,13 @@ public class BaseCollisionChecker {
 	 *         coordinates with respect to the line segment formed by the first
 	 *         two specified coordinates.
 	 */
-	public static int relativeCCW(final double pX1, final double pY1, double pX2, double pY2, double pPX, double pPY) {
-		// TODO Float instead of double!!! (All test will still run?)
+	public static int relativeCCW(final float pX1, final float pY1, float pX2, float pY2, float pPX, float pPY) {
 		pX2 -= pX1;
 		pY2 -= pY1;
 		pPX -= pX1;
 		pPY -= pY1;
-		double ccw = pPX * pY2 - pPY * pX2;
-		if (ccw == 0.0) {
+		float ccw = pPX * pY2 - pPY * pX2;
+		if (ccw == 0.0f) {
 			// The point is colinear, classify based on which side of
 			// the segment the point falls on. We can calculate a
 			// relative value using the projection of PX,PY onto the
@@ -93,7 +92,7 @@ public class BaseCollisionChecker {
 			// outside of the segment in the direction of the particular
 			// endpoint used as the origin for the projection.
 			ccw = pPX * pX2 + pPY * pY2;
-			if (ccw > 0.0) {
+			if (ccw > 0.0f) {
 				// Reverse the projection to be relative to the original X2,Y2
 				// X2 and Y2 are simply negated.
 				// PX and PY need to have (X2 - X1) or (Y2 - Y1) subtracted
@@ -104,12 +103,12 @@ public class BaseCollisionChecker {
 				pPX -= pX2;
 				pPY -= pY2;
 				ccw = pPX * pX2 + pPY * pY2;
-				if (ccw < 0.0) {
-					ccw = 0.0;
+				if (ccw < 0.0f) {
+					ccw = 0.0f;
 				}
 			}
 		}
-		return (ccw < 0.0) ? -1 : ((ccw > 0.0) ? 1 : 0);
+		return (ccw < 0.0f) ? -1 : ((ccw > 0.0f) ? 1 : 0);
 	}
 
 	// ===========================================================
